requirements.txt
fastapi==0.116.1  
uvicorn==0.35.0  
sqlalchemy==2.0.43  
pandas==2.3.2  
openpyxl==3.1.5  
streamlit==1.49.1  
requests==2.32.5  
python-dotenv==1.1.1  
langchain==0.3.27  
langchain-core==0.3.76  
langchain-google-genai==2.0.10  
langgraph==0.6.7  
langgraph-checkpoint==2.1.1  
google-generativeai==0.8.5  
google-ai-generativelanguage==0.6.15  
pydantic==2.11.9  
python-multipart==0.0.20  
email-validator==2.0.0 
